Refactor del código. Distribuido en diferentes archivos.
