+++ dibujar.go
@@ -1,0 +1,97 @@
+package main
+
+import "fmt"
+
+/* Dibuja las manos acordes con las jugadas */
+func dibujar(maquina int, usuario int, nombreUsuario1 string, nombreUsuario2 string) {
+	if usuario == 1 && maquina == 1 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    _______            " + "            _______    \n" +
+			"---'   ____)           " + "           (____   '---\n" +
+			"      (_____)          " + "          (_____)      \n" +
+			"      (_____)          " + "          (_____)      \n" +
+			"      (____)           " + "           (____)      \n" +
+			"---.__(___)            " + "            (___)__.---\n\n")
+	}
+	if usuario == 1 && maquina == 2 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    _______            " + "           ________    \n" +
+			"---'   ____)           " + "        __(____    '---\n" +
+			"      (_____)          " + "      _(_______        \n" +
+			"      (_____)          " + "     (_________        \n" +
+			"      (____)           " + "      (________        \n" +
+			"---.__(___)            " + "         (_________.---\n\n")
+	}
+	if usuario == 1 && maquina == 3 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    _______            " + "           ________    \n" +
+			"---'   ____)           " + "      ____(____    '---\n" +
+			"      (_____)          " + "     (_________        \n" +
+			"      (_____)          " + "     (_________        \n" +
+			"      (____)           " + "           (____)      \n" +
+			"---.__(___)            " + "            (___)__.---\n\n")
+	}
+	if usuario == 2 && maquina == 1 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "            _______    \n" +
+			"---'    ____)__        " + "           (____   '---\n" +
+			"        _______)_      " + "          (_____)      \n" +
+			"        _________)     " + "          (_____)      \n" +
+			"        ________)      " + "           (____)      \n" +
+			"---._________)         " + "            (___)__.---\n\n")
+	}
+	if usuario == 2 && maquina == 2 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "           ________    \n" +
+			"---'    ____)__        " + "        __(____    '---\n" +
+			"        _______)_      " + "      _(_______        \n" +
+			"        _________)     " + "     (_________        \n" +
+			"        ________)      " + "      (________        \n" +
+			"---._________)         " + "         (_________.---\n\n")
+	}
+	if usuario == 2 && maquina == 3 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "           ________    \n" +
+			"---'    ____)__        " + "      ____(____    '---\n" +
+			"        _______)_      " + "     (_________        \n" +
+			"        _________)     " + "     (_________        \n" +
+			"        ________)      " + "           (____)      \n" +
+			"---._________)         " + "            (___)__.---\n\n")
+	}
+	if usuario == 3 && maquina == 1 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "            _______    \n" +
+			"---'    ____)____      " + "           (____   '---\n" +
+			"        _________)     " + "          (_____)      \n" +
+			"        _________)     " + "          (_____)      \n" +
+			"      (____)           " + "           (____)      \n" +
+			"---.__(___)            " + "            (___)__.---\n\n")
+	}
+	if usuario == 3 && maquina == 2 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "           ________    \n" +
+			"---'    ____)____      " + "        __(____    '---\n" +
+			"        _________)     " + "      _(_______        \n" +
+			"        _________)     " + "     (_________        \n" +
+			"      (____)           " + "      (________        \n" +
+			"---.__(___)            " + "         (_________.---\n\n")
+	}
+	if usuario == 3 && maquina == 3 {
+		fmt.Printf("" +
+			nombreUsuario2 + "      " + "                  " + nombreUsuario1 + "\n" +
+			"    ________           " + "           ________    \n" +
+			"---'    ____)____      " + "      ____(____    '---\n" +
+			"        _________)     " + "     (_________        \n" +
+			"        _________)     " + "     (_________        \n" +
+			"      (____)           " + "           (____)      \n" +
+			"---.__(___)            " + "            (___)__.---\n\n")
+	}
+}
